+++ src/advent_of_code/y2022/day01.zig
@@ -5,14 +5,14 @@ const Result = root.Result;
 pub const Solution = root.Solution{ .year = 2022, .day = 1, .run = run };
 
 fn run(input: []const u8) !Result {
-    var max = [_]i32{0} ** 3;
+    var max = [_]u32{0} ** 3;
     var chunks = std.mem.split(u8, input, "\n\n");
     while (chunks.next()) |chunk| {
-        var sum: i32 = 0;
+        var sum: u32 = 0;
         var lines = std.mem.split(u8, chunk, "\n");
         while (lines.next()) |line| {
             if (line.len > 0) {
-                sum += try std.fmt.parseInt(i32, line, 10);
+                sum += try std.fmt.parseInt(u32, line, 10);
             }
         }
 
@@ -28,7 +28,7 @@ fn run(input: []const u8) !Result {
     return Result.of(max[0], max[0] + max[1] + max[2]);
 }
 
-fn compareFn(context: void, a: i32, b: i32) std.math.Order {
+fn compareFn(context: void, a: u32, b: u32) std.math.Order {
     _ = context;
     return std.math.order(b, a);
 }
@@ -52,6 +52,6 @@ test "2022.01 sample" {
     ;
 
     var res = try run(data);
-    try std.testing.expectEqual(@as(i32, 24000), res.part1.int);
-    try std.testing.expectEqual(@as(i32, 45000), res.part2.int);
+    try std.testing.expectEqual(@as(u64, 24000), res.part1.int);
+    try std.testing.expectEqual(@as(u64, 45000), res.part2.int);
 }
